Uh this was a weeks worth of work, I probably shouldve steadily committed but Ill try to explain what developed over the course.
I added a lot of the implementations that were discussed in the last meeting,

Dosage Count: I added the ability to input a dosage count and dosage per bottle but currently there is no tracking system currently it's just variables stored as an integer with no effect upon hitting taken or not taken as we still need to implement a form of a "Day Cycle" where the app automatically resets at night for the next day to then resupplu notifications etc.

I added Emergency contact number, currently just stored as a number and retained on the edit information.

Changed the Medication Time to be input through TimePicker which is an embedded android package that allows ease of selecting a time.

Condensed the days on the AddMedication form so that it doesn't take up as much space as it did.

Added a Text field for Warning Label Directions in case there are any concerns that the user wants to jot down so that they can be reminded of it when opening the medication screen again.

Added a divider on the Main dashboard screen so that elements in the list are separated to give clarity to each individual medication.

Added 2 png files that serve as the Arrows for the drop down menu on the dashboard screen.

Added the Edit Medication button which allows the user to revisit all the information provided initially on the AddMedication screen as well as make necessary changes.

Added Edit Medication to the android manifest so that it can be detected when opened, otherwise crashes the app because of no detectable path.

Updated MedicationAdpater, DashboardActivity, MedicationViewModel, MedicationDao, MedicationRepository, MedicationDatabase to all reflect the added stored variables so that Android's Room extension through SQLite can track the added information persistently.

Created the editmedication layout as well as adjusted existing layouts to reflect the added information as well.

I believe that is all.