Fixed my Dosage counter method so that it can properly decrement by 1 now when it is checked, however it came with me removing the logic that prevented it from being checked more than once because I kept running into errors where locking out the checkbox when then cause the checkbox to no longer work and display properly. Something with the communication between the DashboardActivity and the MedicationAdapter caused this interference.
I also added better logic for sorting the medications so now instead of just being displayed as AM to PM it is now sorted by relevance to the current day and which medication is then next up according to their respective times and days.

I also made it to where the notifications aren't making alarms for days that were in the past and then pushing instantly because initially when it was scheduling an alarm it would also schedule one for that current week so if you made an alarm for monday at 3 PM on a friday it would also send that alarm right away because the previous monday of the week had passed so the alarm would push, but now it looks for the next valid day in the future instead of just within the week and beyond.